Added Error Handling for Empty Source Conns
If at any time a source isn't connected to anything, a new warning
message will pop up letting the user know that it isn't being used.
Added as a warning because it might be intentional. Also, added another
call of checkModel at the end of the program just to reprint this any
any future warning messages.